+++ scripts/build-plugin/build.js
@@ -14,7 +14,6 @@ module.exports = (api ,options = {}) => {
   const { commandArgs, rootDir } = context;
   const getItemOutputPath = (name, dir = 'lib/') => dir + name + '/';
   let taskList = [];
-  let beforeTaskList = [];
   let entry;
   let output;
   let pkgInfo;
@@ -43,14 +42,6 @@ module.exports = (api ,options = {}) => {
       apiInfo = {...defaultApiInfo, unNeedSplit: false};
 
       buildMain(rootDir, sourceMap);
-      // beforeTaskList.push(async (ctx, next) => {
-      //   await initPkg(entry, pkgInfo, output, sourceMap, apiInfo, isMain);
-      // })
-      // taskList.push(async function(ctx, next) {
-      //   await initPkg(entry, pkgInfo, output, sourceMap, apiInfo, isMain);
-      //   gulpRelease(api, { entry, output, isMain });
-      //   await next();
-      // });
     } else {
       entry = path.resolve(rootDir, sourceMap[apiName].path);
       apiInfo = {...defaultApiInfo, ...sourceMap[apiName]};
@@ -59,20 +50,10 @@ module.exports = (api ,options = {}) => {
         sourceMap[apiName].pkgInfo.forEach(i => {
           pkgInfo = i;
           output = outputDir + getItemOutputPath(i.name, sourceMap[apiName].outputDir);
-          // taskList.push(async function(ctx, next) {
-          //   await initPkg(entry, pkgInfo, output, sourceMap, apiInfo, isMain);
-          //   gulpRelease(api, { entry, output, isMain });
-          //   await next();
-          // });
         });
       } else {
         pkgInfo = sourceMap[apiName].pkgInfo;
         output = outputDir + getItemOutputPath(sourceMap[apiName].pkgInfo.name, sourceMap[apiName].outputDir);
-        // taskList.push(async function(ctx, next) {
-        //   await initPkg(entry, pkgInfo, output, sourceMap, apiInfo, isMain);
-        //   gulpRelease(api, { entry, output, isMain });
-        //   await next();
-        // });
       }
     }
     
@@ -88,11 +69,6 @@ module.exports = (api ,options = {}) => {
       unNeedSplit: options.unNeedSplit,
     };
     sourceMap = null;
-    // taskList.push(async function(ctx, next) {
-    //     await initPkg(entry, pkgInfo, output, sourceMap, apiInfo, isMain);
-    //     gulpRelease(api, { entry, output, isMain });
-    //     await next();
-    // });
   }
   rm.sync(path.resolve(rootDir, output));
   webpackRelease(api, {
@@ -109,16 +85,6 @@ module.exports = (api ,options = {}) => {
     await initPkg(entry, pkgInfo, output, sourceMap, apiInfo, isMain);
   });
   onHook('after.build.compile', () => {
-    // if(!isMain) {
-      gulpRelease(api, { entry, output, isMain });
-    // } else {
-    //   rollupRelease(entry, pkgInfo, output, sourceMap, apiInfo, isMain)
-    // }
-    
-    // compose(taskList)().then(function() {
-    //   // logger('END', {status: 'SUCCESS'});
-    // }).catch(function(err) {
-    //   console.log(err);
-    // });
+    gulpRelease(api, { entry, output, isMain });
   });
 }